Converter files copying, data files update from master (#294)
* Fix for Imperator::Character::getCulture() causing crashes (#256)

* Fix wrong code name (#257) #patch

* A few tests (#230)

* Added simple Imp Character tests

* Fix typo in ProvinceTests

* Update tests' CMakeLists

* Update CMakeLists.txt

* Update CMakeLists.txt

* Update CMakeLists.txt

* Update test.yml

* Update tests' cmakelists

* Update test.yml

* Update test.yml

* Update test.yml

* Update test.yml

* organize test files

* add temp line to workflow

* Update test.yml

* Fix for unlinked family members crashing conversion (#264) #patch

* fix for unlinked family members crashing conversion

* Update Family.cpp

* Fix for crashes apparently caused by recent data files PR, new faith icons by Plati (#268) #patch

* Revert "I:R centuries support - Part 1 (#251)"

This reverts commit 4fa3df31be193ba911efee557367d21a180b8d94.

* remove unused folder

* Faith icons by Plati

* Converter files copying

+++ ImperatorToCK3/Outputter/WorldOutputter.cs
@@ -1,23 +1,20 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.IO;
 using ImperatorToCK3.CK3;
 using commonItems;
 
 namespace ImperatorToCK3.Outputter {
 	public static class WorldOutputter {
 		public static void OutputWorld(World ck3World, Configuration theConfiguration) {
-			Logger.Info("Clearing the output mod folder.");
 			var directoryToClear = "output/" + theConfiguration.OutputModName;
-			System.IO.DirectoryInfo di = new DirectoryInfo(directoryToClear);
-			foreach (FileInfo file in di.EnumerateFiles()) {
-				file.Delete();
-			}
-			foreach (DirectoryInfo dir in di.EnumerateDirectories()) {
-				dir.Delete(true);
+			var di = new DirectoryInfo(directoryToClear);
+			if (di.Exists) {
+				Logger.Info("Clearing the output mod folder.");
+				foreach (FileInfo file in di.EnumerateFiles()) {
+					file.Delete();
+				}
+				foreach (DirectoryInfo dir in di.EnumerateDirectories()) {
+					dir.Delete(true);
+				}
 			}
 
 			var outputName = theConfiguration.OutputModName;
@@ -67,7 +64,7 @@ namespace ImperatorToCK3.Outputter {
 
 		private static void CreateModFolder(string outputName) {
 			var modPath = Path.Combine("output", outputName);
-			Directory.CreateDirectory(modPath);
+			SystemUtils.TryCreateFolder(modPath);
 		}
 
 		private static void CreateFolders(string outputName) {
